requirements.txt
alembic==1.13.2
altgraph==0.17.4
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.3.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
cachelib==0.13.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cryptography==43.0.3
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
duckdb==1.1.3
email_validator==2.2.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flaskwebgui==1.1.3
fonttools==4.42.1
fqdn==1.5.1
greenlet==3.0.3
idna==3.4
infinity==1.5
intervals==0.9.2
ipykernel==6.25.1
ipython==8.15.0
isoduration==20.11.0
itsdangerous==2.2.0
JayDeBeApi==1.2.3
jedi==0.19.0
Jinja2==3.1.2
JPype1==1.5.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
msal==1.31.1
msgspec==0.19.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
paramiko==3.5.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
Pygments==2.16.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.10.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==8.0
rapidfuzz==3.2.0
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
Send2Trash==1.8.2
six==1.16.0
smartsheet-dataframe==0.2.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
sshtunnel==0.4.0
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
validators==0.33.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==3.0.3
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
